+++ protocol-designer/src/localization/en/protocol_overview.json
@@ -1,3 +1,5 @@
 {
-  "protocol_overview": "Protocol overview"
+  "protocol_overview": "Protocol overview",
+  "edit": "Edit",
+  "add_labware": "Add labware"
 }
