+++ protocol-designer/src/pages/ProtocolOverview/DeckSetup/DeckSetup.module.css
@@ -1,0 +1,36 @@
+.slot_overlay {
+  position: absolute;
+
+  /* from legacy --absolute-fill */
+  top: 0;
+
+  /* from legacy --absolute-fill */
+  right: 0;
+
+  /* from legacy --absolute-fill */
+  bottom: 0;
+
+  /* from legacy --absolute-fill */
+  left: 0;
+
+  /* from legacy --absolute-fill */
+
+  z-index: 1;
+  padding: 0.5rem;
+  background-color: color-mod(var(--c-black) alpha(0.75));
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: flex-start;
+  color: white;
+  font-size: var(--fs-body-1);
+  border-radius: 0.5rem;
+}
+
+.appear_on_mouseover {
+    opacity: 0;
+  
+    &:hover {
+      opacity: 1;
+    }
+  }
